storage-node: cli: Split commands to separate files.
