+++ storage-node/packages/cli/package.json
@@ -44,6 +44,7 @@
     "axios": "^0.19.2",
     "chalk": "^2.4.2",
     "lodash": "^4.17.11",
-    "meow": "^5.0.0"
+    "meow": "^5.0.0",
+    "ipfs-only-hash": "^1.0.2"
   }
 }
